Add files via upload
In this project I worked with a fake advertising data set, indicating whether or not a particular internet user clicked on an Advertisement on a company website. I tried to create a model that will predict whether or not they will click on an ad based off the features of that user.I used a jupyter notebook to create the model.